face_recognition/requirements.txt
# SpoqPlus 안면인식 시스템 - Python 라이브러리
# 설치 명령어: pip install -r requirements.txt

# 핵심 AI/ML 라이브러리
face_recognition==1.3.0
opencv-python==4.8.1.78
mediapipe==0.10.0
numpy==1.24.3
scikit-learn==1.3.0
Pillow==10.0.1
dlib==19.24.2

# 웹 서버 및 API
flask==2.3.3
flask-cors>=4.0.0
requests==2.31.0

# 데이터베이스
mysql-connector-python>=8.0.33
pymysql==1.1.0

# 유틸리티
python-dotenv==1.0.0
schedule==1.2.0

# 이미지/비디오 처리
imageio==2.31.5
imageio-ffmpeg==0.4.9

# 로깅 및 모니터링
colorama==0.4.6

# 개발/테스트 도구 (선택사항)
pytest==7.4.3
pytest-cov==4.1.0

# 추가 보안 기능을 위한 패키지
scipy==1.11.0

# MediaPipe 기반 얼굴 인식 서비스 필수 패키지
mediapipe==0.10.21
opencv-python==4.10.0.84
flask==3.0.3
flask-cors==5.0.0
numpy==2.1.3
requests==2.32.3
python-dotenv==1.0.1

# 이미지 처리 및 분석
Pillow==10.4.0
scikit-image==0.24.0

# 데이터베이스 연결
PyMySQL==1.1.1
SQLAlchemy==2.0.36

# 보안 및 암호화
cryptography==43.0.3
bcrypt==4.2.1

# 유틸리티
python-dateutil==2.9.0.post0
pytz==2024.2

# 로깅 및 모니터링
colorlog==6.9.0

# 테스트
pytest==8.3.4
pytest-cov==6.0.0

# HTTP 클라이언트
httpx==0.28.1

# JSON 처리
orjson==3.10.12

# 환경 설정
pydantic==2.10.3
pydantic-settings==2.7.0

# 비동기 처리
asyncio==3.4.3

# 시스템 모니터링
psutil==6.1.0

# 캐싱
redis==5.2.1

# 웹 서버
gunicorn==23.0.0
waitress==3.0.2 